renaming default_pars as pars
