+++ src/laser_polio/__init__.py
@@ -1,5 +1,5 @@
 from .distributions import *  # noqa F403
-from .default_pars import *  # noqa F403
+from .pars import *  # noqa F403
 from .model import *  # noqa F403
 from .utils import *  # noqa F403
